refactor: move Support files to their own folder
